requirements.txt
absl-py==2.1.0
accelerate==0.18.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
anyio==3.7.1
apex==0.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.4
coloredlogs==15.0.1
conda==22.9.0
conda-build==3.21.9
conda-content-trust==0+unknown
conda-package-handling==1.8.1
crcmod==1.7
cryptography==37.0.1
cycler==0.11.0
datasets==2.13.2
decorator==5.1.1
deepspeed==0.7.3
diffusers==0.14.0
dill==0.3.6
docopt==0.6.2
einops==0.6.1
exceptiongroup==1.1.2
fastapi==0.100.1
filelock==3.6.0
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
fvcore==0.1.6
glob2==0.7
h11==0.14.0
hjson==3.1.0
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.3
imageio==2.22.3
importlib-metadata==4.11.3
iopath==0.1.10
ipython==7.31.1
jax==0.1.37
jedi==0.18.1
Jinja2==2.10.1
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.4
libarchive-c==2.9
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.0.1
mat4py==0.6.0
matplotlib==3.5.3
matplotlib-inline==0.1.2
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.3
mmpretrain==1.2.0
mmsegmentation==1.2.2
model-index==0.1.11
modelindex==0.0.2
mpi4py==3.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.14
mypy-extensions==0.4.3
ninja==1.10.2.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnx==1.13.1
onnxruntime==1.14.1
opencd==1.1.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.1.2
pipreqs==0.4.13
pkginfo==1.8.2
platformdirs==2.5.2
portalocker==2.7.0
prettytable==3.7.0
prompt-toolkit==3.0.20
protobuf==3.20.3
psutil==5.8.0
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyarrow==12.0.1
pycocotools==2.0.6
pycosat==0.6.3
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.2
Pygments==2.17.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyre-extensions==0.0.23
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.7.1
ruamel-yaml-conda==0.15.100
safetensors==0.3.0
scikit-learn==1.0.2
scipy==1.7.3
setuptools==60.2.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
starlette==0.27.0
supervision==0.3.2
sympy==1.10.1
tabulate==0.9.0
termcolor==2.2.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
timm==0.6.13
tokenizers==0.13.3
tomli==2.0.1
toolz==0.11.2
torch==1.12.1
torchtext==0.13.1
torchvision==0.13.1
tqdm==4.65.2
traitlets==5.1.1
transformers==4.27.4
typed-ast==1.4.3
typing==3.7.4.3
typing_extensions==4.7.1
typing-inspect==0.8.0
urllib3==1.26.8
uvicorn==0.22.0
wcwidth==0.2.5
wheel==0.37.1
xxhash==3.4.1
yacs==0.1.8
yapf==0.32.0
yarg==0.1.9
yarl==1.9.4
zipp==3.8.0
